Seems to work but not catching all of the user IDs for replacement.

+++ slackex2html.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 # Create array of usernames to IDs
 # Create cache of User avatars
 # Get number of entries in file with ".|length"
@@ -6,11 +6,28 @@
 # use jq to extract each element of indexed item and assign to bash var
 # output HTML using generated vars
 
+
+# Check message for UID pattern <U..........>. Use sed.. Then set variable for user with AWK against 
+# cleaned up list of users to store actual user's real name in variable. Then use sed again 
+# to replace it in the message. 
+# To then repeat this until there are no longer any user IDs in the message. Maybe this 
+# Should be a function?  
+# Use " while ! grep UID.Patter msg; do" for iteration ...
+
 # This for the user to configure.
-DEBUG="0"
+DEBUG="1"
 SRCDIR="/Users/scott/Downloads/CatalystSlackExport" # Set this if you dont' want to use the current working directory.
 DSTDIR="/Users/scott/SlackTest"
 mkdir -p $DSTDIR
+# This is where we start. It should be the root of the archive from Slack.
+if [ -z $SRCDIR ]; then
+	SRCDIR=$(pwd)
+fi
+
+# Pre run Cleanup
+if [ -f slackex2html.debug]; then 
+	rm $SRCDIR/slackex2html.debug
+fi
 
 
 # Check to see if we are running on MacOS or Linux, then do some setup.
@@ -40,18 +57,15 @@ fi
 # Setup a debug function..
 Debug() {
 	if [ $DEBUG -eq 1 ]; then
-		echo -e "debug: $1"
+		echo -e "debug:\n$1" | tee -a  $SRCDIR/slackex2html.debug
 		fi
 	}
 
-# This is where we start. It should be the root of the archive from Slack.
-if [ -z $SRCDIR ]; then
-	SRCDIR=$(pwd)
-fi
+
 
 # First we need to create the destination structure.
 cd $SRCDIR
-SRCDIRS=$(find . -type d | sed "s/^\.\///")
+SRCDIRS=$(find . -mindepth 1 -type d | sed "s/^\.\///" | sort )
 
 # then create the new structure in the destination
 cd $DSTDIR
@@ -59,18 +73,35 @@ for DIR in $SRCDIRS; do
 	mkdir -p $DIR 2> /dev/null
 done
 
+# Create an associative array for the user name to ID mapping.
+Debug " $SRCDIR/users.json"
+URECORDS=$( jq ".|length" $SRCDIR/users.json)
+INDEX=0
+declare -A Users
+echo "Reading Users in to array for replacements later"
+while [ $INDEX -lt $URECORDS ]; do
+        ID=$( jq -r ".[$INDEX].id" $SRCDIR/users.json )
+        NAME=$( jq -r ".[$INDEX].real_name" $SRCDIR/users.json )
+        Debug "$ID,$NAME"
+        INDEX=$( expr $INDEX + 1 )
+        Users[$ID]="$NAME"
+done
+
 # Process each file. The final dir structure will mirror the archive's structure, 
 # but include sub folders for the year under each channel folder. 
 # Files in each folder will be by month. 
 #
+echo "Now processing Channels"
 for DIR in $SRCDIRS; do 
-	Debug "DIR $DIR"
+	echo "Channel #$DIR"
+	
 	cd ${SRCDIR}/${DIR}
 	Debug "PWD $(pwd)"
-	for MONTH in 01 02 03 04 05 06 07 08 09 10 11 12; do		
-		
-		# Setup the initial Table.
+	for MONTH in 01 02 03 04 05 06 07 08 09 10 11 12; do	
+		Debug "MONTH $MONTH"		
+
 		for INPUT in $( ls ????-$MONTH-??.json 2> /dev/null| sort); do
+			Debug "INPUT at Start -$INPUT-"
 			DATE=$( echo $INPUT | sed "s/.json$//")
 			YEAR=$(echo $DATE | sed "s/\(^....\)-.*$/\1/")
 			mkdir -p "${DSTDIR}/${DIR}/${YEAR}"
@@ -79,6 +110,7 @@ for DIR in $SRCDIRS; do
 				echo "<html><body>" > $OUTPUT
 			fi
 			Debug "MONTH $MONTH\n YEAR $YEAR\n OUTPUT $OUTPUT\n PWD $(pwd)"
+			# Setup the initial Table.
 			echo "<h1>$DATE</h1><table>" >> $OUTPUT
 			RECORDS=$(jq '.|length' $INPUT)   # File record count (number of messages in the file)
 			LASTUSER="null"
@@ -94,6 +126,21 @@ for DIR in $SRCDIRS; do
 					# this should be fixed to make this an anchor tag.
 					MSG=$( echo $MSG | tr '<|>' ' ')
 				fi
+				# Replace the UIDs with actual user names on mentions
+			
+				while echo "$MSG" | grep -sq "<U..........>" ; do 
+					ID=$(echo $MSG | sed "s/^.*\(<U..........>\).*$/\1/;s/[<>]//g")
+					MSG=$(echo $MSG | sed "s/<${ID}>/${Users[$ID]}>/g")		
+					Debug "Found $ID in:\n$MSG"
+				done 
+
+				# Replace the UIDs with actual user names on @ mentions
+			
+				while echo "$MSG" | grep -sq "<@U..........>" ; do
+					ID=$(echo $MSG | sed "s/^.*\(<@U..........>\).*$/\1/;s/[<>@]//g")
+					MSG=$(echo $MSG | sed "s/<@${ID}>/@${Users[$ID]}/g")		
+					Debug "Found $ID in:\n$MSG"
+				done 
 				# This sets up the Username and avatar, but sets them to null if the last
 				# entry was from the same user. This gives a cleaner look in the rendered html
 				USER=$( jq -r ".[$INDEX]|.user_profile.real_name" $INPUT)
@@ -105,18 +152,23 @@ for DIR in $SRCDIRS; do
 					AVATAR=""
 				fi
 				LASTUSER=$USER
-
 				Debug "$TS\n$USER\n$MSG\n$AVATAR"
 				# Create a new row in the table;
 				echo "<tr><td><b>${USERNAME}</b></td><td><img src="$AVATAR"></td><td>$TS:</td><td> $MSG</td></tr>" >> $OUTPUT
 				# increase the INDEX
 				INDEX=$( expr $INDEX + 1)
-	
-
 			done 
-			echo "</table>" >> $OUTPUT
-		done  
-		echo "</body></html>" >> $OUTPUT
+			Debug "Close table INPUT $INPUT OUTPUT $OUTPUT"
+			if [ -f $OUTPUT ]; then
+				echo "</table>" >> $OUTPUT
+			fi	
+		done 
+		Debug "Close tag INPUT $INPUT \n OUTPUT $OUTPUT" 
+		# We first check to see if the file exists as it may not.
+		# This happens when there is no source file for the months in this loop.
+		if [ -f "$OUTPUT" ]; then
+			echo "</body></html>" >> $OUTPUT
+		fi
 	done
 
 done
